+++ src/main/java/com/nowcoder/community/service/LikeService.java
@@ -15,14 +15,7 @@ public class LikeService {
     private RedisTemplate redisTemplate;
 
     // 点赞
-    public void like(int userId, int entityType, int entityId, int entityUserId){
-//        String entityLikeKey = RedisKeyUtil.getEntityLikeKey(entityType, entityId);
-//        boolean isMember = redisTemplate.opsForSet().isMember(entityLikeKey, userId);
-//        if(isMember){
-//            redisTemplate.opsForSet().remove(entityId, userId);
-//        } else{
-//            redisTemplate.opsForSet().add(entityLikeKey, entityId);
-//        }
+    public void like(int userId, int entityType, int entityId, int entityUserId) {
         redisTemplate.execute(new SessionCallback() {
             @Override
             public Object execute(RedisOperations operations) throws DataAccessException {
@@ -30,12 +23,13 @@ public class LikeService {
                 String userLikeKey = RedisKeyUtil.getUserLikeKey(entityUserId);
 
                 boolean isMember = operations.opsForSet().isMember(entityLikeKey, userId);
+
                 operations.multi();
 
-                if(isMember){
+                if (isMember) {
                     operations.opsForSet().remove(entityLikeKey, userId);
                     operations.opsForValue().decrement(userLikeKey);
-                }else{
+                } else {
                     operations.opsForSet().add(entityLikeKey, userId);
                     operations.opsForValue().increment(userLikeKey);
                 }
@@ -46,19 +40,19 @@ public class LikeService {
     }
 
     // 查询某实体点赞的数量
-    public long findEntityLikeCount(int entityType, int entityId){
+    public long findEntityLikeCount(int entityType, int entityId) {
         String entityLikeKey = RedisKeyUtil.getEntityLikeKey(entityType, entityId);
         return redisTemplate.opsForSet().size(entityLikeKey);
     }
 
     // 查询某人对某实体的点赞状态
-    public long findEntityLikeStatus(int userId, int entityType, int entityId){
+    public int findEntityLikeStatus(int userId, int entityType, int entityId) {
         String entityLikeKey = RedisKeyUtil.getEntityLikeKey(entityType, entityId);
         return redisTemplate.opsForSet().isMember(entityLikeKey, userId) ? 1 : 0;
     }
 
     // 查询某个用户获得的赞
-    public int findUserLikeCount(int userId){
+    public int findUserLikeCount(int userId) {
         String userLikeKey = RedisKeyUtil.getUserLikeKey(userId);
         Integer count = (Integer) redisTemplate.opsForValue().get(userLikeKey);
         return count == null ? 0 : count.intValue();
